Fixed failing test 2026 Aug 19 23.08
[ci]
